pyproject.toml
[project]
name = "webspace-sync"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kugeleis", email = "9948777+Kugeleis@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-box>=7.4.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
webspace-sync = "webspace_sync:main"

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "bump-my-version>=1.3.0",
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]

[tool.bumpversion]
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = true
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
pre_commit_hooks = ["uv sync", "git add uv.lock"]

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'
regex = false
